# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: JornadaMS API
  version: 1.0.0-mvp1
  description: API REST privada do MVP-1 do JornadaMS.
servers:
  - url: http://localhost:8000
    description: Desenvolvimento local
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /api/v1/employees:
    get:
      tags: [Employees]
      operationId: listEmployees
      x-mvp: MVP-1
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/EmployeeStatus'
      responses:
        '200':
          description: Lista de colaboradores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [Employees]
      operationId: createEmployee
      x-mvp: MVP-1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreate'
      responses:
        '201':
          description: Colaborador criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/employees/{employee_id}:
    parameters:
      - $ref: '#/components/parameters/EmployeeId'
    get:
      tags: [Employees]
      operationId: getEmployee
      x-mvp: MVP-1
      responses:
        '200':
          description: Colaborador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Employees]
      operationId: updateEmployee
      x-mvp: MVP-1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdate'
      responses:
        '200':
          description: Colaborador atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/employees/{employee_id}/activate:
    parameters:
      - $ref: '#/components/parameters/EmployeeId'
    post:
      tags: [Employees]
      operationId: activateEmployee
      x-mvp: MVP-1
      responses:
        '200':
          description: Colaborador ativado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/employees/{employee_id}/deactivate:
    parameters:
      - $ref: '#/components/parameters/EmployeeId'
    post:
      tags: [Employees]
      operationId: deactivateEmployee
      x-mvp: MVP-1
      responses:
        '200':
          description: Colaborador desativado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: jornada_access
  parameters:
...
    EmployeeId:
      in: path
      name: employee_id
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: page_size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflito de estado ou idempotência
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Dados semanticamente inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    EmployeeStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
...
    EmployeeCreate:
      type: object
      required: [name, registration_code, branch_id]
      properties:
        name:
          type: string
          minLength: 1
        registration_code:
          type: string
          minLength: 1
        punch_identifier:
          type: [string, 'null']
        branch_id:
          type: string
          format: uuid
        department_id:
          type: [string, 'null']
          format: uuid
        position_id:
          type: [string, 'null']
          format: uuid
        user_id:
          type: [string, 'null']
          format: uuid
    EmployeeUpdate:
      type: object
      description: Campos opcionais para atualização parcial do colaborador.
      properties:
        name:
          type: string
          minLength: 1
        registration_code:
          type: string
          minLength: 1
        punch_identifier:
          type: [string, 'null']
        branch_id:
          type: string
          format: uuid
        department_id:
          type: [string, 'null']
          format: uuid
        position_id:
          type: [string, 'null']
          format: uuid
        user_id:
          type: [string, 'null']
          format: uuid
    Employee:
      allOf:
        - $ref: '#/components/schemas/EmployeeCreate'
        - type: object
          required: [id, company_id, status]
          properties:
            id:
              type: string
              format: uuid
            company_id:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/EmployeeStatus'
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlation_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items: {}
            correlation_id:
              type: string
    EmployeePage:
      type: object
      required: [items, page, page_size, total_items, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
...